% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\alias{new_coordinates}
\alias{coordinates.default}
\alias{coordinates.Spatial}
\title{A coordinates tuple with vctrs}
\usage{
new_coordinates(x = double(), y = double(), crs = character())

coordinates(x = double(), y = double(), crs = character())

\method{coordinates}{default}(x = double(), y = double(),
  crs = character())

\method{coordinates}{Spatial}(x = double(), y = double(),
  crs = character())
}
\arguments{
\item{x}{x stuff}

\item{y}{y stuff}

\item{proj}{stuff}
}
\value{
vctrs_coordinates
}
\description{
A coordinates tuple with vctrs
}
\examples{

new_coordinates()
coordinates()
## note how this must be explicitly double (not integer) but the
## easy to use constructor handles the coercion
x <- new_coordinates(rnorm(10), as.double(1:10))
x <- coordinates(rnorm(10), 1:10)
x
as.data.frame(x)
plot(x)
}
