% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\title{Create coordinate object}
\usage{
coords(x, y)
}
\arguments{
\item{x}{as one or a list of numeric vectors of finite length}

\item{y}{as one or a list of numeric vectors of finite length}
}
\value{
coordinate object containing `$x` and `$y` of class `coords`.
}
\description{
Create coordinate object
}
