% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_coords.R
\name{edge}
\alias{edge}
\title{determine edge coordinates}
\usage{
edge(obj)
}
\arguments{
\item{coords}{object}
}
\value{
min and max values of coordinate pairs
}
\description{
determine edge coordinates
}
