% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_coords.R
\name{length.coords}
\alias{length.coords}
\title{Print length of coordinate pairs (method)}
\usage{
\method{length}{coords}(obj)
}
\arguments{
\item{coords}{object}
}
\value{
length of coordinate pairs
}
\description{
Print length of coordinate pairs (method)
}
