% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_coords.R
\name{Math.vcoords}
\alias{Math.vcoords}
\title{group method for all standard mathematical transformations}
\usage{
\method{Math}{vcoords}(x)
}
\arguments{
\item{vcoords}{object}
}
\value{
coordinate pairs and transformed values
}
\description{
group method for all standard mathematical transformations
}
