% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{new_coordinates}
\alias{new_coordinates}
\title{Title}
\usage{
new_coordinates(x = double(), y = double(), crs = character())
}
\arguments{
\item{y}{}
}
\description{
Title
}
\examples{

new_coordinates()
coordinates()
x <- new_coordinates(rnorm(10))
x
as.data.frame(x)
}
