% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithm_vcoords.R
\name{Ops.vcoords}
\alias{Ops.vcoords}
\title{simple arithmetic methods for vcoords objects}
\usage{
\method{Ops}{vcoords}(obj1, obj2)
}
\arguments{
\item{obj1}{vcoords object or vector one}

\item{obj2}{vcoords object or vector two}
}
\value{
coordinate pairs and transformed values
}
\description{
simple arithmetic methods for vcoords objects
}
\note{
allows value transformation in the format e.g. `(obj1*obj2)` or `(obj1*2)`
}
