% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_coords.R
\name{print.coords}
\alias{print.coords}
\title{Print coordinate pairs (method)}
\usage{
\method{print}{coords}(obj)
}
\arguments{
\item{coords}{object}
}
\value{
coordinate pairs `(x, y)`.
}
\description{
Print coordinate pairs (method)
}
\note{
the `print` function is executed implicitly when a result is displayed.
}
