% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vcoords.R
\name{print.vcoords}
\alias{print.vcoords}
\title{Print coordinate pairs (method)}
\usage{
\method{print}{vcoords}(obj)
}
\arguments{
\item{coords}{object}
}
\value{
coordinate pairs `(x, y; v)`.
}
\description{
Print coordinate pairs (method)
}
\note{
the `print` function is executed implicitly when a result is displayed.
}
