% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithm_vcoords.R
\name{test_location}
\alias{test_location}
\title{Test if coordinates are the same in both objects}
\usage{
test_location(obj1, obj2)
}
\arguments{
\item{obj1}{vcoords object one}

\item{obj2}{vcoords object two}
}
\value{
boolean (TRUE or FALSE)
}
\description{
Test if coordinates are the same in both objects
}
