% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcoords.R
\name{vcoords}
\alias{vcoords}
\title{Create coordinate and value paired object}
\usage{
vcoords(x, y, v)
}
\arguments{
\item{x}{as one or a list of numeric vectors of finite length}

\item{y}{as one or a list of numeric vectors of finite length}

\item{v}{value attached to coordinate}
}
\value{
coordinate object containing `$x`, `$y` and `$v` of class `coords` and `vcoords`.
}
\description{
Create coordinate and value paired object
}
