% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xycoords.R
\name{xcoords}
\alias{xcoords}
\title{extract xcoordinates}
\usage{
xcoords(obj)
}
\arguments{
\item{coords}{object containing x and y coordinates}
}
\value{
list of x coords
}
\description{
extract xcoordinates
}
