% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xycoords.R
\name{ycoords}
\alias{ycoords}
\title{extract ycoordinates}
\usage{
ycoords(obj)
}
\arguments{
\item{coords}{object containing x and y coordinates}
}
\value{
list of y coords
}
\description{
extract ycoordinates
}
